Detect libass <8c8741f if unsigned int has at most 34 value bits
It is impossible to distinguish between an old libass with a 35-bit-wide
unsigned int and a new libass without invoking undefined behavior.

If sscanf("17179869184", "%d", &unsigned_value) was well-defined,
the width of unsigned int could be detected and hence new libass
could be distinguished from old libass if the width was small.