Put lines with quirky timestamps in stream header in FFmpeg and VLC
This also affects mpv, which uses FFmpeg to parse external subtitle files.

This will break if/when FFmpeg and VLC fix their parsers.

This also works for current mkvmerge (so patching is no longer
required), but again, this will break when its parser is fixed,
so the old mechanism with the [end of codec private] marker is kept.