Reduce \blur to speed up VSFilterMod & old VSFilter rendering
Let the old-Aegisub messages have 1.5x line spacing so that their
3 lines match the new-Aegisub message's 4 lines, and avoid hiding
and re-printing the first and last line of the message, which are
exactly the same in all cases.

Avoid blurring the hanging punctuation CWords, as there's no point.

(\be wasn't slow, ironically, but adjust it similarly.)

Reduce the \blur strength. Of course, the output
becomes a little more hazy, but this isn't obvious.

The time to render a frame in sorayuki's current VSFilterMod
has dropped from about 25 seconds (!) to about 2.5 for me.