Make 200% sure libass 3bbd058 message isn't visible on later frames
Some players may treat the video as if it has huge duration due to the
large timestamps the subtitles contain. For example, Perian uses 9:59:59:59;
I can imagine some other player doing the same. It's better to show nothing
after some time than to show wrong information, so make sure the libass
3bbd058-580ae21 message can be visible only during the first 10 seconds.
The current code is not enough in libass 7913e4a+, which clips the timestamp
to the range of int32, which is smaller than the full duration of the event.