Really put quirky-timestamp lines in stream header in FFmpeg & VLC
Commit fd0b5f3a2e097d556d8df884711f02cab5e7ff77 attempted to do this,
but I had misunderstood the FFmpeg and VLC code (and not tested enough).
They use sscanf(" "), which skips spaces but does not require them,
so omitting the space had no effect. Instead, add a garbage character
to one of the timestamps that has no effect in libass and Aegisub (and
even mkvmerge) but breaks parsing in FFmpeg and VLC fail, as their full
sscanf formats require a comma in the same position.